requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.25
pydantic==2.6.1
python-dotenv==1.0.1
cos-python-sdk-v5==1.9.30
